balloon-game-mobile/src/app/+leaderboard/leaderboard.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { LeaderboardComponent } from './leaderboard.component';

describe('Component: Leaderboard', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [LeaderboardComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([LeaderboardComponent],
      (component: LeaderboardComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(LeaderboardComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(LeaderboardComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-leaderboard></app-leaderboard>
  `,
  directives: [LeaderboardComponent]
})
class LeaderboardComponentTestController {
}

